pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "podcast-updates"
version = "0.1.0"
description = "Daily podcast aggregation pipeline — ingest, transcribe, and cross-analyze news podcasts"
requires-python = ">=3.11"
dependencies = [
    "feedparser>=6.0",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "anthropic>=0.40",
    "jinja2>=3.1",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
moonshine = ["moonshine-voice"]
whisper = ["faster-whisper>=1.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[project.scripts]
podcast-updates = "src.pipeline:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
